more fixes

+++ eole/trainer.py
@@ -265,12 +265,12 @@ class Trainer(object):
 
     def _update_average(self, step):
         if self.moving_average is None:
-            copy_params = [params.detach().float() for params in self.model.parameters()]
+            copy_params = [params.detach() for params in self.model.parameters()]
             self.moving_average = copy_params
         else:
             average_decay = max(self.average_decay, 1 - (step + 1) / (step + 10))
             for (i, avg), cpt in zip(enumerate(self.moving_average), self.model.parameters()):
-                self.moving_average[i] = (1 - average_decay) * avg + cpt.detach().float() * average_decay
+                self.moving_average[i] = (1 - average_decay) * avg + cpt.detach() * average_decay
 
     def train(
         self,
